More robust detection of output to session.
It turns out that the log vector can actually shrink or remain the same size,
even though something is output to the session. To handle this, we compare the
entire log before and after executing the expression. Pretty wasteful, but it's
quick enough in my testing.

The other problem is knowing what the difference between the new and old log is.
This is currently handled completely ad hoc, by searching for the last 10 lines
from the old log in the new one, and assuming everything after the last match
is new output. This is sort of slow, but reasonably robust, unless you output
the same thing all the time, but in that case the log tends to change size.