Add ability to edit a function at a particular line.
This is invoked the same way as in Dyalog, when the name is followed by a line
number inside brackets.

When triggering Emacs from inside the Dyalog editor, this doesn't work, because
there is no way to get information on anything but the name the cursor is on. It
works from the Dyalog session and from within Emacs however.