Add new indentation system, and enable it by default.
The new indentation code handles logical lines (lines separated by diamonds)
correctly, and is more strict when looking for what keywords to indent to, e.g.
a :Else will look for a :If, :Select or a :Trap, not any keyword that causes
indentation (like the old code did).

Note that indent-region still is extremely slow, since it searches backward for
how to indent each line of the region (e.g it's O(n^2)).