Fix indentation of labels.
This fixes indentation of labels that was buggy in several cases previously:

Labels in tradfuns inside namespaces were note indented correctly before.

Labels can now appear before keywords, without breaking font lock or
indentation.

The code after the label is now indented correctly, e.g. there is no space after
the label, unless the following token is a keyword or comment (and indentation
of comments is disabled).

Code after a label that is supposed to be indented (e.g. it is inside an :If) is
now indented correctly even if the label is longer than the desired indentation.

The new forward indentation for regions still doesn't indent labels correctly.