Update next-line, and initialize indent-info correctly.
Since indent-line-with can insert or remove spaces at the beginning of the line,
we need to use a marker that is updated so we move to the new logical line end
position.

And indent-info should be initialized with the contents of the previous line,
otherwise we will get errors when the first line of the region attempts to
reduce indentation (e.g. it is an :Else or an :EndIf).