Make behaviour of beginning-of-defun more intuitive.
Now moving to the beginning of a defun moves to the beginning of the defun point
is in, instead of the immediately preceding defun. For example, if point is in a
tradfn, but there is a local dfun definition inside the tradfn before point,
point will move to the tradfn definition. If point is inside the local dfn
on the other hand, it will move to the start of the dfn.

This is similar to how up-list works in Emacs.