Significantly improve imenu performance for files with many functions.
The improved performance comes from two changes:

Avoiding searching backwards for the function definition start, this is
pointless since we move forward through the buffer to the next function
definition when creating the imenu index, so we now we are at the start of a
function.

The second part of the improvement comes from optimizing moving to the end of a
tradfn definition by using code specialized for tradfns. This avoids a bunch of
calls to syntax-ppss.