Make dyalog-in-dfun faster by using the syntax-ppss cache.
syntax-ppss uses a cache to avoid constantly re-parsing the buffer from the
beginning. However, when we use a different syntax table the cache isn't
invalidated correctly, causing errors. For some reason, this only happens during
jit-lock. Previously we worked around these errors by detecting when syntax-ppss
gave an erronous result and falling back to scan-lists.

However, this caused us to re-parse from the beginning of the buffer every time
in the worst case. When font-locking at the end of large buffers this would make
e.g. scrolling very slow.

So we change the algorithm for detecting if we are inside a dfn. We only use
syntax-ppss, but manually skip regular parens, and only count braces. This might
require several calls to syntax-ppss, but every call should hit the cache and
thus be fast.

This change completely removes dyalog-in-dfun from the profile when scrolling
through large buffers.