Speed up fontification of dfuns inside tradfuns.
We know we are at the start of a defun after calling dyalog-next-defun, so there
is no need to call dyalog-defun-info just to check if we reached a dfun, we can
just check if we are looking at a open brace. Also, by passing on the info that
we are at the start of a defun to dyalog-defun-info, we can skip a bunch of
processing to parse out defun info.