Fix whitespace for string literals.
There is always a space between string literals and numbers, but not between
string literals and names.

Note that the algorithm in dyalog-fix-whitespace should be fixed to look at the
type of token that is next to a punctuation character, instead of just what
character it is. For example, a name can end in a digit, but is falsely
classified as a number currently.